cleanup cookie & commerce helpers
