+++ scripts/commerce.js
@@ -1,41 +1,8 @@
-/**
- * Get a cookie
- * @param cname the name of the cookie
- */
-export function getCookie(cname) {
-  let value = decodeURIComponent(
-    // eslint-disable-next-line prefer-template
-    document.cookie.replace(new RegExp('(?:(?:^|.*;)\\s*' + encodeURIComponent(cname).replace(/[\\-\\.\\+\\*]/g, '\\$&') + '\\s*\\=\\s*([^;]*).*$)|^.*$'), '$1'),
-  ) || null;
-  if (value && ((value.substring(0, 1) === '{' && value.substring(value.length - 1, value.length) === '}') || (value.substring(0, 1) === '[' && value.substring(value.length - 1, value.length) === ']'))) {
-    try {
-      value = JSON.parse(value);
-    } catch (e) {
-      return value;
-    }
-  }
-  return value;
-}
+// eslint-disable-next-line import/no-cycle
+import { getCookie } from './scripts.js';
 
-/**
-* Set the content of a cookie
-* @param {string} cname The cookie name (or property)
-* @param {string} cvalue The cookie value
-* @param {number} expTime The cookie expiry time (default 30 days)
-* @param {string} path The cookie path (optional)
-*
-*/
-export function setCookie(cname, cvalue, expTime = 30 * 1000 * 60 * 60 * 24, path = '/') {
-  const today = new Date();
-  today.setTime(today.getTime() + (expTime));
-  const expires = 'expires='.concat(today.toGMTString());
-  const cookieString = cname.concat('=')
-    .concat(cvalue)
-    .concat(';')
-    .concat(expires)
-    .concat(';path=')
-    .concat(path);
-  document.cookie = cookieString; // cname + '=' + cvalue + ';' + expires + ';path=' + path;
+export function getCommerceBase() {
+  return window.DanaherConfig !== undefined ? window.DanaherConfig.intershopDomain + window.DanaherConfig.intershopPath : 'https://shop.lifesciences.danaher.com/INTERSHOP/rest/WFS/DANAHERLS-LSIG-Site/-';
 }
 
 /**
